pyproject.toml
[project]
name = "workout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langflow>=1.0.19.post2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.1",
]

[project.scripts]
main = "ai:main"

[tool.uv]
package = true
